Create getMatchedVariation util
