+++ plugins/woocommerce/client/blocks/assets/js/base/stores/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './does-cart-item-match-attributes';
+export * from './get-matched-variation';
