+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/variation-selector/frontend.ts
@@ -3,8 +3,10 @@
  */
 import { store, getContext } from '@wordpress/interactivity';
 import { SelectedAttributes } from '@woocommerce/stores/woocommerce/cart';
+import { getMatchedVariation } from '@woocommerce/stores/utils';
 import type { ChangeEvent } from 'react';
 import type { ProductDataStore } from '@woocommerce/stores/woocommerce/product-data';
+import type { AvailableVariation } from '@woocommerce/stores/utils';
 
 /**
  * Internal dependencies
@@ -12,7 +14,6 @@ import type { ProductDataStore } from '@woocommerce/stores/woocommerce/product-d
 import type {
 	AddToCartWithOptionsStore,
 	Context as AddToCartWithOptionsStoreContext,
-	AvailableVariation,
 } from '../frontend';
 import setStyles from './set-styles';
 
@@ -118,43 +119,6 @@ const isAttributeValueValid = ( {
 		).length;
 
 		return matchingAttributes >= attributesToMatch;
-	} );
-};
-
-const getMatchedVariation = (
-	availableVariations: AvailableVariation[],
-	selectedAttributes: SelectedAttributes[]
-) => {
-	if (
-		! Array.isArray( availableVariations ) ||
-		! Array.isArray( selectedAttributes ) ||
-		availableVariations.length === 0 ||
-		selectedAttributes.length === 0
-	) {
-		return null;
-	}
-	return availableVariations.find( ( availableVariation ) => {
-		return Object.entries( availableVariation.attributes ).every(
-			( [ attributeName, attributeValue ] ) => {
-				const attributeMatched = selectedAttributes.some(
-					( variationAttribute ) => {
-						const isSameAttribute =
-							variationAttribute.attribute === attributeName;
-						if ( ! isSameAttribute ) {
-							return false;
-						}
-
-						return (
-							variationAttribute.value === attributeValue ||
-							( variationAttribute.value &&
-								attributeValue === '' )
-						);
-					}
-				);
-
-				return attributeMatched;
-			}
-		);
 	} );
 };
